Stage texture database + wall materials
I added several textures I made during my internship free to use in the project.  Also two wall materials one for the building and one for the church are now inside the project. Also we have an cliff texture that potentially can be used. Delete the textures if there not needed to we save space